requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
supabase==1.0.3
httpx>=0.23.0,<0.25.0 